1 out of 4

+++ Mirza_Lorena_practic/Controllers/MoviesController.cs
@@ -20,9 +20,47 @@ namespace Mirza_Lorena_practic.Controllers
         }
 
         // GET: Movies
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(string sortOrder,string currentFilter,string searchString,int? pageNumber)
         {
-            return View(await _context.Movies.ToListAsync());
+            ViewData["CurrentSort"] = sortOrder;
+            ViewData["TitleSortParm"] = String.IsNullOrEmpty(sortOrder) ? "title_desc" : "";
+            ViewData["PriceSortParm"] = sortOrder == "Price" ? "price_desc" : "Price";
+
+            if (searchString != null)
+            {
+                pageNumber = 1;
+            }
+            else
+            {
+                searchString = currentFilter;
+            }
+
+            ViewData["CurrentFilter"] = searchString;
+            var movies = from b in _context.Movies
+                        select b;
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                movies = movies.Where(s => s.Title.Contains(searchString));
+            }
+            switch (sortOrder)
+            {
+                case "title_desc":
+                    movies = movies.OrderByDescending(b => b.Title);
+                    break;
+                case "Price":
+                    movies = movies.OrderBy(b => b.Price);
+                    break;
+                case "price_desc":
+                    movies = movies.OrderByDescending(b => b.Price);
+                    break;
+                default:
+                    movies = movies.OrderBy(b => b.Title);
+                    break;
+            }
+
+            int pageSize = 2;
+            return View(await PaginatedList<Movie>.CreateAsync(movies.AsNoTracking(), pageNumber ??
+           1, pageSize));
         }
 
         // GET: Movies/Details/5
@@ -34,7 +72,11 @@ namespace Mirza_Lorena_practic.Controllers
             }
 
             var movie = await _context.Movies
-                .FirstOrDefaultAsync(m => m.ID == id);
+             .Include(s => s.Orders)
+             .ThenInclude(e => e.Customer)
+             .AsNoTracking()
+             .FirstOrDefaultAsync(m => m.ID == id);
+
             if (movie == null)
             {
                 return NotFound();
@@ -54,13 +96,22 @@ namespace Mirza_Lorena_practic.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,Title,Director,Price")] Movie movie)
+        public async Task<IActionResult> Create([Bind("Title,Director,Price")] Movie movie)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(movie);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(movie);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            catch (DbUpdateException /* ex*/)
+            {
+
+                ModelState.AddModelError("", "Unable to save changes. " +
+                "Try again, and if the problem persists ");
             }
             return View(movie);
         }
@@ -84,51 +135,53 @@ namespace Mirza_Lorena_practic.Controllers
         // POST: Movies/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,Title,Director,Price")] Movie movie)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != movie.ID)
+            if (id == null)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            var studentToUpdate = await _context.Movies.FirstOrDefaultAsync(s => s.ID == id);
+            if (await TryUpdateModelAsync<Movie>(
+            studentToUpdate,
+            "",
+            s => s.Director, s => s.Title, s => s.Price))
             {
                 try
                 {
-                    _context.Update(movie);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /* ex */)
                 {
-                    if (!MovieExists(movie.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(movie);
+            return View(studentToUpdate);
         }
 
         // GET: Movies/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
         {
             if (id == null)
             {
                 return NotFound();
             }
 
-            var movie = await _context.Movies
+            var movie = await _context.Movies.AsNoTracking()
                 .FirstOrDefaultAsync(m => m.ID == id);
             if (movie == null)
             {
                 return NotFound();
+            }
+
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["ErrorMessage"] =
+                "Delete failed. Try again";
             }
 
             return View(movie);
@@ -140,9 +193,23 @@ namespace Mirza_Lorena_practic.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var movie = await _context.Movies.FindAsync(id);
-            _context.Movies.Remove(movie);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+
+            if (movie == null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.Movies.Remove(movie);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (DbUpdateException /* ex */)
+            {
+
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
         private bool MovieExists(int id)
